app/requirements.dev.txt
pylint
pylint-django
autopep8
